figma-transformer: extract text normalizer
