+++ figma-transformer/src/Scenegraph/ScenegraphNormalizer.php
@@ -11,11 +11,15 @@ final class ScenegraphNormalizer
 {
     private const GEOMETRY_SEMANTICS_COMPONENT_SOURCE_CLONE = 'component_source_clone';
 
+    private readonly TextNormalizer $textNormalizer;
+
     public function __construct(
         private readonly ScenegraphIndex $index = new ScenegraphIndex(),
         private readonly VectorGeometryNormalizer $vectorGeometryNormalizer = new VectorGeometryNormalizer(),
-        private readonly PaintNormalizer $paintNormalizer = new PaintNormalizer()
+        private readonly PaintNormalizer $paintNormalizer = new PaintNormalizer(),
+        ?TextNormalizer $textNormalizer = null
     ) {
+        $this->textNormalizer = $textNormalizer ?? new TextNormalizer($this->vectorGeometryNormalizer);
     }
 
     /**
@@ -397,7 +401,7 @@ final class ScenegraphNormalizer
         }
 
         if ( 'TEXT' === $type ) {
-            $text = $this->normalizeText($node, $blobs, $id, $diagnostics);
+            $text = $this->textNormalizer->normalizeText($node, $blobs, $id, $diagnostics);
             if ( ! empty($text) ) {
                 $node['figma_text'] = $text;
             }
@@ -2166,656 +2170,6 @@ final class ScenegraphNormalizer
     }
 
     /**
-     * @param array<string, mixed> $node
-     * @return array<string, mixed>
-     */
-    private function normalizeText(array $node, array $blobs = array(), string $nodeId = '', array &$diagnostics = array()): array
-    {
-        $text = array();
-
-        foreach ( array('characters', 'text') as $key ) {
-            if ( isset($node[$key]) && is_scalar($node[$key]) ) {
-                $text['characters'] = (string) $node[$key];
-                break;
-            }
-        }
-
-        if ( ! isset($text['characters']) && isset($node['textData']['characters']) && is_scalar($node['textData']['characters']) ) {
-            $text['characters'] = (string) $node['textData']['characters'];
-        }
-
-        $style = array();
-        if ( is_array($node['style'] ?? null) ) {
-            $style = $this->normalizeTextStyle($node['style']);
-        }
-
-        $rootStyle = $this->normalizeTextStyle($node);
-        foreach ( $rootStyle as $key => $value ) {
-            if ( ! array_key_exists($key, $style) ) {
-                $style[$key] = $value;
-            }
-        }
-
-        if ( ! empty($style) ) {
-            $text['style'] = $style;
-        }
-
-        if ( isset($text['characters']) && is_scalar($text['characters']) ) {
-            $iconFallback = $this->fontAwesomeIconNameFallback((string) $text['characters'], $style);
-            if ( null !== $iconFallback ) {
-                $text['icon_name'] = (string) $text['characters'];
-                $text['characters'] = $iconFallback;
-            }
-        }
-
-        $derivedLayout = $this->normalizeDerivedTextLayout($node, $blobs, $nodeId, $diagnostics);
-        if ( ! empty($derivedLayout) ) {
-            $text['derived_layout'] = $derivedLayout;
-        }
-
-        $segments = $this->normalizeStyledTextSegments($node);
-        if ( ! empty($segments) ) {
-            $text['segments'] = $segments;
-        }
-
-        return $text;
-    }
-
-    /**
-     * @param array<string, mixed> $style
-     */
-    private function fontAwesomeIconNameFallback(string $characters, array $style): ?string
-    {
-        $fontFamily = strtolower((string) ($style['font_family'] ?? ''));
-        $postscript = strtolower((string) ($style['font_postscript_name'] ?? ''));
-        if ( ! str_contains($fontFamily, 'font awesome') && ! str_contains($postscript, 'fontawesome') ) {
-            return null;
-        }
-
-        return match ( strtolower(trim($characters)) ) {
-            'sparkle', 'sparkles' => '✦',
-            'circle-check', 'check-circle' => '✓',
-            'circle', 'circle-small' => '●',
-            'arrow-right' => '→',
-            'arrow-left' => '←',
-            'arrow-up' => '↑',
-            'arrow-down' => '↓',
-            'chevron-right' => '›',
-            'chevron-left' => '‹',
-            default => null,
-        };
-    }
-
-    /**
-     * @param array<string, mixed> $node
-     * @return array<string, mixed>
-     */
-    private function normalizeDerivedTextLayout(array $node, array $blobs = array(), string $nodeId = '', array &$diagnostics = array()): array
-    {
-        $source = is_array($node['derivedTextData'] ?? null) ? $node['derivedTextData'] : array();
-        if ( empty($source) ) {
-            return array();
-        }
-
-        $layout = array();
-        if ( is_array($source['layoutSize'] ?? null) ) {
-            $size = array();
-            foreach ( array('x' => 'width', 'y' => 'height', 'width' => 'width', 'height' => 'height') as $sourceKey => $targetKey ) {
-                if ( ! isset($size[$targetKey]) && isset($source['layoutSize'][$sourceKey]) && is_numeric($source['layoutSize'][$sourceKey]) ) {
-                    $size[$targetKey] = (float) $source['layoutSize'][$sourceKey];
-                }
-            }
-            if ( ! empty($size) ) {
-                $layout['size'] = $size;
-            }
-        }
-
-        if ( is_array($source['baselines'] ?? null) ) {
-            $layout['baseline_count'] = count($source['baselines']);
-            $baselines = array();
-            foreach ( $source['baselines'] as $baseline ) {
-                if ( ! is_array($baseline) ) {
-                    continue;
-                }
-                $normalized = array();
-                foreach ( array('width', 'lineY', 'lineHeight', 'lineAscent', 'firstCharacter', 'endCharacter') as $key ) {
-                    if ( isset($baseline[$key]) && is_numeric($baseline[$key]) ) {
-                        $normalized[$key] = (float) $baseline[$key];
-                    }
-                }
-                if ( is_array($baseline['position'] ?? null) ) {
-                    foreach ( array('x', 'y') as $axis ) {
-                        if ( isset($baseline['position'][$axis]) && is_numeric($baseline['position'][$axis]) ) {
-                            $normalized['position_' . $axis] = (float) $baseline['position'][$axis];
-                        }
-                    }
-                }
-                if ( ! empty($normalized) ) {
-                    $baselines[] = $normalized;
-                }
-            }
-            if ( ! empty($baselines) ) {
-                $layout['baselines'] = $baselines;
-            }
-        }
-
-        if ( is_array($source['glyphs'] ?? null) ) {
-            $layout['glyph_count'] = count($source['glyphs']);
-            $glyphPaths = array();
-            $characters = isset($node['textData']['characters']) && is_scalar($node['textData']['characters']) ? (string) $node['textData']['characters'] : ( isset($node['characters']) && is_scalar($node['characters']) ? (string) $node['characters'] : '' );
-            $characterList = '' !== $characters ? preg_split('//u', $characters, -1, PREG_SPLIT_NO_EMPTY) : array();
-            if ( ! is_array($characterList) ) {
-                $characterList = array();
-            }
-            foreach ( $source['glyphs'] as $index => $glyph ) {
-                if ( ! is_array($glyph) ) {
-                    continue;
-                }
-
-                $glyphPath = array();
-                foreach ( array('x', 'y', 'advance', 'fontSize', 'fontIndex', 'firstCharacter', 'endCharacter') as $key ) {
-                    if ( isset($glyph[$key]) && is_numeric($glyph[$key]) ) {
-                        $glyphPath[$key] = (float) $glyph[$key];
-                    }
-                }
-                if ( isset($glyph['firstCharacter']) && is_numeric($glyph['firstCharacter']) && isset($characterList[(int) $glyph['firstCharacter']]) ) {
-                    $glyphPath['character'] = $characterList[(int) $glyph['firstCharacter']];
-                }
-                if ( is_array($glyph['position'] ?? null) ) {
-                    foreach ( array('x', 'y') as $axis ) {
-                        if ( isset($glyph['position'][$axis]) && is_numeric($glyph['position'][$axis]) ) {
-                            $glyphPath['position_' . $axis] = (float) $glyph['position'][$axis];
-                        }
-                    }
-                }
-
-                if ( isset($glyph['commandsBlob']) ) {
-                    $bytes = $this->vectorGeometryNormalizer->readCommandBlobBytes($glyph['commandsBlob'], $blobs);
-                    if ( null !== $bytes ) {
-                        $decoded = $this->vectorGeometryNormalizer->classifyVectorCommandBlob($bytes);
-                        if ( 'path' === $decoded['status'] ) {
-                            $glyphPath['data'] = $decoded['path'];
-                        } elseif ( 'unsupported' === $decoded['status'] ) {
-                            $diagnostics[] = array(
-                                'severity' => 'warning',
-                                'code'     => 'unsupported_text_glyph_command_blob',
-                                'message'  => 'Unsupported Figma text glyph command blob was omitted from derived glyph metadata.',
-                                'context'  => array('node_id' => $nodeId, 'glyph_index' => $index),
-                            );
-                        }
-                        // 'empty' blobs (e.g. whitespace glyphs encoded as a single
-                        // 0x00 byte) carry no drawable outline and are not warnings.
-                    }
-                }
-
-                if ( empty($glyphPath) ) {
-                    continue;
-                }
-                $glyphPaths[] = $glyphPath;
-            }
-            if ( ! empty($glyphPaths) ) {
-                $layout['glyph_paths'] = $glyphPaths;
-            }
-        }
-        if ( is_array($source['fontMetaData'] ?? null) ) {
-            $fonts = array();
-            foreach ( $source['fontMetaData'] as $font ) {
-                if ( ! is_array($font) ) {
-                    continue;
-                }
-                $fonts[] = array(
-                    'family' => (string) ($font['key']['family'] ?? ''),
-                    'style' => (string) ($font['key']['style'] ?? ''),
-                    'font_weight' => isset($font['fontWeight']) && is_numeric($font['fontWeight']) ? (int) $font['fontWeight'] : null,
-                    'font_line_height' => isset($font['fontLineHeight']) && is_numeric($font['fontLineHeight']) ? (float) $font['fontLineHeight'] : null,
-                );
-            }
-            if ( ! empty($fonts) ) {
-                $layout['fonts'] = $fonts;
-            }
-        }
-
-        return $layout;
-    }
-
-    /**
-     * @param array<string, mixed> $source
-     * @return array<string, mixed>
-     */
-    private function normalizeTextStyle(array $source): array
-    {
-        $style = array();
-
-        foreach ( array(
-            'fontFamily' => 'font_family',
-            'fontPostScriptName' => 'font_postscript_name',
-            'fontWeight' => 'font_weight',
-            'textAlignHorizontal' => 'text_align_horizontal',
-            'textAlignVertical' => 'text_align_vertical',
-            'textDecoration' => 'text_decoration',
-        ) as $sourceKey => $targetKey ) {
-            if ( isset($source[$sourceKey]) && is_scalar($source[$sourceKey]) && '' !== (string) $source[$sourceKey] ) {
-                $style[$targetKey] = (string) $source[$sourceKey];
-            }
-        }
-
-        if ( isset($source['fontName']) && is_array($source['fontName']) ) {
-            if ( isset($source['fontName']['family']) && is_scalar($source['fontName']['family']) ) {
-                $style['font_family'] = (string) $source['fontName']['family'];
-            }
-            if ( isset($source['fontName']['postscript']) && is_scalar($source['fontName']['postscript']) ) {
-                $style['font_postscript_name'] = (string) $source['fontName']['postscript'];
-            }
-            if ( ! isset($style['font_weight']) && isset($source['fontName']['style']) && is_scalar($source['fontName']['style']) ) {
-                $fontWeight = $this->fontWeightFromStyle((string) $source['fontName']['style']);
-                if ( null !== $fontWeight ) {
-                    $style['font_weight'] = $fontWeight;
-                }
-            }
-        }
-
-        foreach ( array('fontSize' => 'font_size', 'lineHeightPx' => 'line_height_px', 'lineHeightPercent' => 'line_height_percent', 'letterSpacing' => 'letter_spacing') as $sourceKey => $targetKey ) {
-            if ( isset($source[$sourceKey]) && is_numeric($source[$sourceKey]) ) {
-                $style[$targetKey] = (float) $source[$sourceKey];
-            }
-        }
-
-        if ( isset($source['lineHeight']) && is_array($source['lineHeight']) && isset($source['lineHeight']['value']) && is_numeric($source['lineHeight']['value']) ) {
-            $lineHeightUnits = strtoupper((string) ($source['lineHeight']['units'] ?? ''));
-            if ( 'PIXELS' === $lineHeightUnits ) {
-                $style['line_height_px'] = (float) $source['lineHeight']['value'];
-            } elseif ( 'RAW' === $lineHeightUnits ) {
-                $style['line_height_raw'] = (float) $source['lineHeight']['value'];
-            } elseif ( str_contains($lineHeightUnits, 'PERCENT') ) {
-                $style['line_height_percent'] = (float) $source['lineHeight']['value'];
-            }
-        }
-
-        if ( isset($source['letterSpacing']) && is_array($source['letterSpacing']) && isset($source['letterSpacing']['value']) && is_numeric($source['letterSpacing']['value']) ) {
-            $letterSpacingUnits = strtoupper((string) ($source['letterSpacing']['units'] ?? 'PIXELS'));
-            if ( 'PIXELS' === $letterSpacingUnits ) {
-                $style['letter_spacing'] = (float) $source['letterSpacing']['value'];
-            } elseif ( 'RAW' === $letterSpacingUnits ) {
-                $style['letter_spacing_em'] = (float) $source['letterSpacing']['value'];
-            } elseif ( str_contains($letterSpacingUnits, 'PERCENT') ) {
-                $style['letter_spacing_em'] = (float) $source['letterSpacing']['value'] / 100;
-            }
-        }
-
-        foreach ( array('color', 'textColor') as $sourceKey ) {
-            $color = $this->normalizeColor($source[$sourceKey] ?? null);
-            if ( null !== $color ) {
-                $style['color'] = $color;
-                break;
-            }
-        }
-
-        foreach ( array('underline' => 'underline', 'strikethrough' => 'strikethrough') as $sourceKey => $targetKey ) {
-            if ( isset($source[$sourceKey]) && is_bool($source[$sourceKey]) ) {
-                $style[$targetKey] = $source[$sourceKey];
-            }
-        }
-
-        // Figma `textCase` enum → CSS text-transform / font-variant. ORIGINAL and
-        // absent values keep the native font casing. SMALL_CAPS_FORCED uppercases
-        // before rendering small caps, matching Figma's render behaviour.
-        if ( isset($source['textCase']) && is_scalar($source['textCase']) && '' !== (string) $source['textCase'] ) {
-            $textCase = strtoupper((string) $source['textCase']);
-            $textTransform = match ( $textCase ) {
-                'UPPER'             => 'uppercase',
-                'LOWER'             => 'lowercase',
-                'TITLE'             => 'capitalize',
-                'SMALL_CAPS_FORCED' => 'uppercase',
-                default             => null,
-            };
-            if ( null !== $textTransform ) {
-                $style['text_transform'] = $textTransform;
-            }
-            if ( 'SMALL_CAPS' === $textCase || 'SMALL_CAPS_FORCED' === $textCase ) {
-                $style['font_variant'] = 'small-caps';
-            }
-        }
-
-        // Figma `paragraphSpacing` (px between paragraphs). The emitter renders a
-        // multi-paragraph text node as a single white-space:pre-line element, so
-        // this is captured for downstream consumers and a diagnostic rather than
-        // emitted as CSS that would not apply.
-        if ( isset($source['paragraphSpacing']) && is_numeric($source['paragraphSpacing']) ) {
-            $style['paragraph_spacing'] = (float) $source['paragraphSpacing'];
-        }
-
-        // Figma `paragraphIndent` (px first-line indent of each paragraph). Decoded
-        // by the Kiwi parser but previously never read here, so it was dropped for
-        // .fig input. Maps directly onto CSS `text-indent` in the emitter.
-        if ( isset($source['paragraphIndent']) && is_numeric($source['paragraphIndent']) ) {
-            $style['paragraph_indent'] = (float) $source['paragraphIndent'];
-        }
-
-        return $style;
-    }
-
-    private function fontWeightFromStyle(string $style): ?int
-    {
-        $style = strtolower(str_replace(array('-', '_'), ' ', $style));
-        if ( str_contains($style, 'thin') ) {
-            return 100;
-        }
-        if ( str_contains($style, 'extra light') || str_contains($style, 'ultra light') ) {
-            return 200;
-        }
-        if ( str_contains($style, 'light') ) {
-            return 300;
-        }
-        if ( str_contains($style, 'regular') || str_contains($style, 'normal') ) {
-            return 400;
-        }
-        if ( str_contains($style, 'medium') ) {
-            return 500;
-        }
-        if ( str_contains($style, 'semi bold') || str_contains($style, 'semibold') || str_contains($style, 'demi bold') ) {
-            return 600;
-        }
-        if ( str_contains($style, 'extra bold') || str_contains($style, 'ultra bold') ) {
-            return 800;
-        }
-        if ( str_contains($style, 'bold') ) {
-            return 700;
-        }
-        if ( str_contains($style, 'black') || str_contains($style, 'heavy') ) {
-            return 900;
-        }
-
-        return null;
-    }
-
-    /**
-     * @param array<string, mixed> $node
-     * @return array<int, array<string, mixed>>
-     */
-    private function normalizeStyledTextSegments(array $node): array
-    {
-        $segments = array();
-        $rawSegments = null;
-        foreach ( array('styledTextSegments', 'segments') as $key ) {
-            if ( is_array($node[$key] ?? null) ) {
-                $rawSegments = $node[$key];
-                break;
-            }
-        }
-
-        if ( ! is_array($rawSegments) ) {
-            // Fall back to character-level override encoding when no segment list is present.
-            return $this->normalizeCharacterStyleOverrideSegments($node);
-        }
-
-        foreach ( $rawSegments as $segment ) {
-            if ( ! is_array($segment) ) {
-                continue;
-            }
-
-            $normalized = array();
-            foreach ( array('characters', 'text') as $key ) {
-                if ( isset($segment[$key]) && is_scalar($segment[$key]) ) {
-                    $normalized['characters'] = (string) $segment[$key];
-                    break;
-                }
-            }
-            foreach ( array('start', 'end') as $key ) {
-                if ( isset($segment[$key]) && is_numeric($segment[$key]) ) {
-                    $normalized[$key] = (int) $segment[$key];
-                }
-            }
-
-            $style = is_array($segment['style'] ?? null) ? $this->normalizeTextStyle($segment['style']) : $this->normalizeTextStyle($segment);
-            if ( ! empty($style) ) {
-                $normalized['style'] = $style;
-            }
-
-            if ( ! empty($normalized) ) {
-                $segments[] = $normalized;
-            }
-        }
-
-        return $segments;
-    }
-
-    /**
-     * Converts the character-level Figma override encoding into the same segment
-     * format produced by {@see normalizeStyledTextSegments}.
-     *
-     * Figma REST API and .fig files expose per-character style overrides via two
-     * parallel fields, with different names and shapes per encoding:
-     *   - REST `characterStyleOverrides` / Kiwi `textData.characterStyleIDs` — one
-     *     integer per character; 0 = base style, N = an entry in the override table.
-     *   - REST `styleOverrideTable` — map from string key (the N above) to a Figma
-     *     style object carrying only the overriding properties.
-     *   - Kiwi `textData.styleOverrideTable` — a `NodeChange[]` where each entry
-     *     carries a `styleID` (the N above) plus the overriding properties
-     *     (`fontName`, `fontSize`, `fillPaints`, ...). It is bridged into the same
-     *     id-keyed map shape the REST path produces.
-     *
-     * Adjacent characters sharing the same override ID are collapsed into a single
-     * run. For each non-base run the override style is compared against the
-     * resolved base style, and only the differing properties (color, font-weight,
-     * etc.) are stored in the segment's `style` key so the emitter emits minimal
-     * `<span>` wrappers.
-     *
-     * @param array<string, mixed> $node
-     * @return array<int, array<string, mixed>>
-     */
-    private function normalizeCharacterStyleOverrideSegments(array $node): array
-    {
-        $textData = is_array($node['textData'] ?? null) ? $node['textData'] : array();
-
-        $overrides = is_array($node['characterStyleOverrides'] ?? null) ? array_values($node['characterStyleOverrides']) : array();
-        if ( empty($overrides) && is_array($textData['characterStyleIDs'] ?? null) ) {
-            $overrides = array_values($textData['characterStyleIDs']);
-        }
-
-        $overrideTable = is_array($node['styleOverrideTable'] ?? null) ? $node['styleOverrideTable'] : array();
-        if ( empty($overrideTable) && is_array($textData['styleOverrideTable'] ?? null) ) {
-            $overrideTable = $this->indexKiwiStyleOverrideTable($textData['styleOverrideTable']);
-        }
-
-        if ( empty($overrides) || empty($overrideTable) ) {
-            return array();
-        }
-
-        // If all override IDs are 0, the entire text uses the base style — no spans needed.
-        $hasNonZero = false;
-        foreach ( $overrides as $id ) {
-            if ( 0 !== (int) $id ) {
-                $hasNonZero = true;
-                break;
-            }
-        }
-        if ( ! $hasNonZero ) {
-            return array();
-        }
-
-        // Resolve the characters string. .fig (Kiwi) nests it under `textData`.
-        $characters = '';
-        foreach ( array('characters', 'text') as $key ) {
-            if ( isset($node[$key]) && is_scalar($node[$key]) ) {
-                $characters = (string) $node[$key];
-                break;
-            }
-        }
-        if ( '' === $characters && isset($textData['characters']) && is_scalar($textData['characters']) ) {
-            $characters = (string) $textData['characters'];
-        }
-        if ( '' === $characters ) {
-            return array();
-        }
-
-        // Build the normalized base style so override deltas can be diffed against it.
-        $baseStyleSource = is_array($node['style'] ?? null) ? $node['style'] : array();
-        $baseStyle = $this->normalizeTextStyle($baseStyleSource);
-        $rootStyle = $this->normalizeTextStyle($node);
-        foreach ( $rootStyle as $key => $value ) {
-            if ( ! array_key_exists($key, $baseStyle) ) {
-                $baseStyle[$key] = $value;
-            }
-        }
-        // Extract text fill color from the base node's fills when not already in
-        // style. REST encodes these as `fills`; .fig (Kiwi) as `fillPaints`.
-        if ( ! isset($baseStyle['color']) ) {
-            $baseFills = $this->firstFillList(array($baseStyleSource['fills'] ?? null, $node['fills'] ?? null, $node['fillPaints'] ?? null));
-            $fillColor = $this->solidFillColor($baseFills);
-            if ( null !== $fillColor ) {
-                $baseStyle['color'] = $fillColor;
-            }
-        }
-
-        // Split into Unicode codepoints (mirrors the approach used elsewhere in this class).
-        $chars = preg_split('//u', $characters, -1, PREG_SPLIT_NO_EMPTY);
-        if ( ! is_array($chars) ) {
-            return array();
-        }
-        $charCount = count($chars);
-
-        // Collapse adjacent characters with the same override ID into runs.
-        $runs = array();
-        $runChars = '';
-        $runId = null;
-
-        for ( $i = 0; $i < $charCount; $i++ ) {
-            $id = isset($overrides[$i]) ? (int) $overrides[$i] : 0;
-            if ( null !== $runId && $id !== $runId ) {
-                $runs[] = array('characters' => $runChars, 'override_id' => $runId);
-                $runChars = '';
-            }
-            $runChars .= $chars[$i];
-            $runId = $id;
-        }
-        if ( '' !== $runChars && null !== $runId ) {
-            $runs[] = array('characters' => $runChars, 'override_id' => $runId);
-        }
-
-        if ( empty($runs) ) {
-            return array();
-        }
-
-        $segments = array();
-        foreach ( $runs as $run ) {
-            $overrideId = (int) $run['override_id'];
-            $segment = array('characters' => $run['characters']);
-
-            if ( 0 !== $overrideId ) {
-                $rawOverride = is_array($overrideTable[(string) $overrideId] ?? null)
-                    ? $overrideTable[(string) $overrideId]
-                    : array();
-
-                if ( ! empty($rawOverride) ) {
-                    $overrideStyle = $this->normalizeTextStyle($rawOverride);
-
-                    // Figma encodes override text color as fills: REST uses `fills`,
-                    // .fig (Kiwi) NodeChange entries use `fillPaints`.
-                    if ( ! isset($overrideStyle['color']) ) {
-                        $overrideFills = $this->firstFillList(array($rawOverride['fills'] ?? null, $rawOverride['fillPaints'] ?? null));
-                        $fillColor = $this->solidFillColor($overrideFills);
-                        if ( null !== $fillColor ) {
-                            $overrideStyle['color'] = $fillColor;
-                        }
-                    }
-
-                    // Keep only properties that differ from the base style.
-                    $delta = array();
-                    foreach ( $overrideStyle as $key => $value ) {
-                        if ( ! array_key_exists($key, $baseStyle) || $baseStyle[$key] !== $value ) {
-                            $delta[$key] = $value;
-                        }
-                    }
-
-                    if ( ! empty($delta) ) {
-                        $segment['style'] = $delta;
-                    }
-                }
-            }
-
-            $segments[] = $segment;
-        }
-
-        return $segments;
-    }
-
-    /**
-     * Bridges the Kiwi `textData.styleOverrideTable` (a `NodeChange[]` where each
-     * entry carries a `styleID`) into the id-keyed map shape that the REST path —
-     * and the rest of {@see normalizeCharacterStyleOverrideSegments} — expects.
-     *
-     * @param array<int|string, mixed> $table
-     * @return array<string, array<string, mixed>>
-     */
-    private function indexKiwiStyleOverrideTable(array $table): array
-    {
-        $indexed = array();
-        foreach ( $table as $entry ) {
-            if ( ! is_array($entry) ) {
-                continue;
-            }
-            if ( ! isset($entry['styleID']) || ! is_numeric($entry['styleID']) ) {
-                continue;
-            }
-            $indexed[(string) (int) $entry['styleID']] = $entry;
-        }
-
-        return $indexed;
-    }
-
-    /**
-     * Returns the first non-empty paint list from a set of candidate sources. Used
-     * to bridge REST `fills` and Kiwi `fillPaints` when resolving text colors.
-     *
-     * @param array<int, mixed> $candidates
-     * @return array<int, mixed>
-     */
-    private function firstFillList(array $candidates): array
-    {
-        foreach ( $candidates as $candidate ) {
-            if ( is_array($candidate) && ! empty($candidate) ) {
-                return $candidate;
-            }
-        }
-
-        return array();
-    }
-
-    /**
-     * Extracts the CSS color from the first visible SOLID fill in a paint list.
-     *
-     * Returns the normalized RGBA array used by the rest of the normalizer, or
-     * null when no usable solid fill is found.
-     *
-     * @param array<int, mixed> $fills
-     * @return array{r: float, g: float, b: float, a?: float}|null
-     */
-    private function solidFillColor(array $fills): ?array
-    {
-        foreach ( $fills as $fill ) {
-            if ( ! is_array($fill) ) {
-                continue;
-            }
-            $type = strtoupper((string) ($fill['type'] ?? 'SOLID'));
-            if ( 'SOLID' !== $type ) {
-                continue;
-            }
-            $color = $this->normalizeColor($fill['color'] ?? null);
-            if ( null === $color ) {
-                continue;
-            }
-            $opacity = isset($fill['opacity']) && is_numeric($fill['opacity']) ? (float) $fill['opacity'] : 1.0;
-            if ( $opacity < 1.0 ) {
-                $color['a'] = $opacity * ($color['a'] ?? 1.0);
-            }
-
-            return $color;
-        }
-
-        return null;
-    }
-
-    /**
      * Capture Figma hyperlink and prototype navigation data so the emitter can produce real anchors.
      *
      * @param array<string, mixed> $node
@@ -2832,7 +2186,7 @@ final class ScenegraphNormalizer
                 }
             }
 
-            $segmentLink = $this->normalizeSegmentHyperlink($node);
+            $segmentLink = $this->textNormalizer->normalizeSegmentHyperlink($node);
             if ( null !== $segmentLink ) {
                 return $segmentLink;
             }
@@ -2884,34 +2238,6 @@ final class ScenegraphNormalizer
         }
         if ( null !== $nodeId ) {
             return array('type' => 'node', 'target_node_id' => $nodeId);
-        }
-
-        return null;
-    }
-
-    /**
-     * @param array<string, mixed> $node
-     * @return array<string, mixed>|null
-     */
-    private function normalizeSegmentHyperlink(array $node): ?array
-    {
-        foreach ( array('styledTextSegments', 'segments') as $key ) {
-            if ( ! is_array($node[$key] ?? null) ) {
-                continue;
-            }
-
-            foreach ( $node[$key] as $segment ) {
-                if ( ! is_array($segment) || ! array_key_exists('hyperlink', $segment) ) {
-                    continue;
-                }
-
-                $link = $this->normalizeHyperlinkValue($segment['hyperlink']);
-                if ( null !== $link ) {
-                    $link['source'] = 'segment';
-                    $link['partial'] = true;
-                    return $link;
-                }
-            }
         }
 
         return null;
